6502Emu: Optimized the CPU disassembler API.
I don't know why I thought it was a good idea, but the disassembler used
to malloc space for each new instruction and free it after printing it,
now it leaves the malloc to the user (main.c) and takes a pointer to the
string buffer, this reduces the amount of malloc's and free's from one
for each instruction to only one for the overridable buffer.

Also, did some clean-up here and there by deleting some debug lines I
left before.